libbpfcontain: remove libbpfcontain and move containerize uprobe logic into rust code
