+++ Cargo.toml
@@ -36,6 +36,7 @@ regex = "1.4.3"
 lazy_static = "1.4.0"
 enum_dispatch = "0.3.5"
 log4rs = "1.0.0"
+procfs = "0.9.1"
 
 [build-dependencies]
 bindgen = "0.56.0"
