+++ bindings.h
@@ -1,2 +1,1 @@
-#include "src/include/libbpfcontain.h"
-#include "src/include/structs.h"
+#include "src/bpf/structs.h"
