--- src/include/libbpfcontain.h
@@ -1,10 +1,0 @@
-// SPDX-License-Identifier: GPL-2.0-or-later
-//
-// BPFContain - Container security with eBPF
-// Copyright (C) 2020  William Findlay
-//
-// Dec. 29, 2020  William Findlay  Created this.
-
-#include "../include/structs.h"
-
-int containerize(u64 policy_id);
